+++ CHANGELOG.md
@@ -13,12 +13,13 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 - Display STR variant filter status on corresponding variantS page
 - Discontinue ClinVar submissions via CSV files and support only submission via API: removed buttons for downloading ClinVar submission objects as CSV files.
+- Warning and reference to Bieseker et al Biesecker when using PP1/BS4 and PP4 in ACMG classifications
+- Warning to not to use PP4 criterion together with PS2/PM6 in ACMG classifications with reference to the SVI Recommendation for _De Novo_ Criteria (PS2 & PM6)
 - Button to directly remove accepted submissions from ClinVar
 ### Fixed
 - Release docs to include instructions for upgrading dependencies
 - Truncated long HGVS descriptions on cancer SNV and SNVs pages
 - Avoid recurrent error by removing variant ranking settings in unranked demo case
-
 
 ## [4.95]
 ### Added